+++ purchases/src/main/kotlin/com/revenuecat/purchases/utils/SerializationException.kt
@@ -1,0 +1,6 @@
+package com.revenuecat.purchases.utils
+
+/**
+ * A generic exception indicating the problem during serialization or deserialization process
+ */
+internal open class SerializationException(message: String, cause: Throwable? = null) : RuntimeException(message, cause)
